darkskywdfunction/requirements.txt
azure-functions
httpx
pytz
